Fixed type conversions for environmental domain ports

+++ libraries/networking/src/DomainHandler.h
@@ -37,7 +37,7 @@ const unsigned short DEFAULT_DOMAIN_SERVER_PORT =
     .contains("HIFI_DOMAIN_SERVER_PORT")
         ? QProcessEnvironment::systemEnvironment()
             .value("HIFI_DOMAIN_SERVER_PORT")
-            .toShort()
+            .toUShort()
         : 40102;
 
 const unsigned short DEFAULT_DOMAIN_SERVER_DTLS_PORT = 
@@ -45,7 +45,7 @@ const unsigned short DEFAULT_DOMAIN_SERVER_DTLS_PORT =
     .contains("HIFI_DOMAIN_SERVER_DTLS_PORT")
         ? QProcessEnvironment::systemEnvironment()
             .value("HIFI_DOMAIN_SERVER_DTLS_PORT")
-            .toShort()
+            .toUShort()
         : 40103;
 
 const quint16 DOMAIN_SERVER_HTTP_PORT = 
@@ -53,7 +53,7 @@ const quint16 DOMAIN_SERVER_HTTP_PORT =
     .contains("HIFI_DOMAIN_SERVER_HTTP_PORT")
         ? QProcessEnvironment::systemEnvironment()
             .value("HIFI_DOMAIN_SERVER_HTTP_PORT")
-            .toShort()
+            .toUInt()
         : 40100;
 
 const quint16 DOMAIN_SERVER_HTTPS_PORT = 
@@ -61,7 +61,7 @@ const quint16 DOMAIN_SERVER_HTTPS_PORT =
     .contains("HIFI_DOMAIN_SERVER_HTTPS_PORT")
         ? QProcessEnvironment::systemEnvironment()
             .value("HIFI_DOMAIN_SERVER_HTTPS_PORT")
-            .toShort()
+            .toUInt()
         : 40101;
 
 const int MAX_SILENT_DOMAIN_SERVER_CHECK_INS = 5;
